requirements.txt
praw
psycopg2
pandas
python-dotenv
